Fix problem with bad setState call.
## Motivation
Prior to this change, we got under certain situtations following:
```
Warning: Cannot update a component (`WidgetFocusProvider`) while
rendering a different component (`Query`). To locate the bad setState()
call inside `Query`, follow the stack trace as described in
https://fb.me/setstate-in-render
    in Query (created by ConnectStoresWrapper[Query] stores=widgets)
    in ConnectStoresWrapper[Query] stores=widgets
    in div (created by Col)
    in Col (created by SearchResult__StyledCol)
    in SearchResult__StyledCol
    in div (created by Row)
    in Row (created by SearchResult__StyledRow)
    in SearchResult__StyledRow
```

## Description
This change will remove the line which causes the problem.
It was meant as a safty net, but not really needed.